RIP commands: regions, autoscale, vm
