+++ internal/command/root/root.go
@@ -14,7 +14,6 @@ import (
 	"github.com/superfly/flyctl/internal/command/agent"
 	"github.com/superfly/flyctl/internal/command/apps"
 	"github.com/superfly/flyctl/internal/command/auth"
-	"github.com/superfly/flyctl/internal/command/autoscale"
 	"github.com/superfly/flyctl/internal/command/certificates"
 	"github.com/superfly/flyctl/internal/command/checks"
 	"github.com/superfly/flyctl/internal/command/config"
@@ -64,7 +63,6 @@ import (
 	"github.com/superfly/flyctl/internal/command/tokens"
 	"github.com/superfly/flyctl/internal/command/turboku"
 	"github.com/superfly/flyctl/internal/command/version"
-	"github.com/superfly/flyctl/internal/command/vm"
 	"github.com/superfly/flyctl/internal/command/volumes"
 	"github.com/superfly/flyctl/internal/command/wireguard"
 	"github.com/superfly/flyctl/internal/flag/flagnames"
@@ -146,9 +144,7 @@ func New() *cobra.Command {
 		resume.New(),     // TODO: deprecate
 		dnsrecords.New(), // TODO: deprecate
 
-		vm.New(),        // TODO: deprecate
-		regions.New(),   // TODO: deprecate
-		autoscale.New(), // TODO: deprecate
+		regions.New(), // TODO: deprecate
 	)
 
 	// if os.Getenv("DEV") != "" {
